+++ Assets/VRM10_Samples/ClothSample/RotateParticle/Runtime/Jobs/Types.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 139cb8a384ad8c64abc4dcfab37bcaca
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
